Moved indexfixtures to test_indexing
