+++ src/snovault/tests/test_indexing.py
@@ -14,9 +14,55 @@ from snovault import (
 )
 
 pytestmark = [pytest.mark.indexing]
-pytest_plugins = [
-    'snovault.tests.indexfixtures',
-]
+
+
+@pytest.fixture(scope='session')
+def app_settings(wsgi_server_host_port, elasticsearch_server, postgresql_server):
+    from .testappfixtures import _app_settings
+    settings = _app_settings.copy()
+    settings['create_tables'] = True
+    settings['persona.audiences'] = 'http://%s:%s' % wsgi_server_host_port
+    settings['elasticsearch.server'] = elasticsearch_server
+    settings['sqlalchemy.url'] = postgresql_server
+    settings['collection_datastore'] = 'elasticsearch'
+    settings['item_datastore'] = 'elasticsearch'
+    settings['indexer'] = True
+    settings['indexer.processes'] = 2
+    return settings
+
+
+@pytest.yield_fixture(scope='session')
+def app(app_settings):
+    from snovault import main
+    app = main({}, **app_settings)
+
+    yield app
+
+    from snovault import DBSESSION
+    DBSession = app.registry[DBSESSION]
+    # Dispose connections so postgres can tear down.
+    DBSession.bind.pool.dispose()
+
+
+@pytest.fixture(autouse=True)
+def teardown(app):
+    import transaction
+    from sqlalchemy import MetaData
+    from zope.sqlalchemy import mark_changed
+    from snovault import DBSESSION
+    from snovault.elasticsearch import create_mapping
+    create_mapping.run(app, skip_indexing=True)
+    session = app.registry[DBSESSION]
+    connection = session.connection().connect()
+    meta = MetaData(bind=session.connection(), reflect=True)
+    for table in meta.sorted_tables:
+        print('Clear table %s' % table)
+        print('Count before -->', str(connection.scalar("SELECT COUNT(*) FROM %s" % table)))
+        connection.execute(table.delete(synchronize_session=False))
+        print('Count after -->', str(connection.scalar("SELECT COUNT(*) FROM %s" % table)), '\n')
+    session.flush()
+    mark_changed(session())
+    transaction.commit()
 
 def test_indexer_queue(app):
     indexer_queue_mirror = app.registry[INDEXER_QUEUE_MIRROR]
